PickUp Anim in + new character model
New character model with the head separated so that we can swap it out later on.
Pick up Animation also in.
Just needs to be added to the animation/blend tree and triggered.
Have moved the previous models into a legacy folder for deprecation and clean up.
Will confirm they are no longer needed before deletion

Rematerialed the new BeanDefaultBody.
Hooked up the animator to the Bean Default Body.

Did notice that test scene does not follow player (or the players current room) but remains locked onto default room.
Have not touched the cameras or the scene (beyond changing out the player)