Column + Tidy up for game folder
I have moved the crypt stuff and materials to a legacy folder.
Renamed characterrigs to characters and placed all the misc items such as colours in there.

I have added the texture to the base model for the egyptian room and the column, meaning they can be dragged straight into the game world, no need to apply mats.

Column is in so we can start getting basic shadow puzzles set up now and get the mechanics implemented and testing.

Simplified the naming of the gameobjects folders (ie one above says egypt instead of everything underneath).